pyproject.toml
[project]
name = "x"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.18",
    "gitpython>=3.1.43",
    "instructor>=1.6.3",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.5",
    "langchain-text-splitters>=0.3.2",
    "numpy>=1.26.4",
    "openai>=1.54.0",
    "psutil>=6.1.0",
    "scikit-learn>=1.5.2",
    "sutil>=0.0.2",
    "tavily-python>=0.5.0",
    "tqdm>=4.66.6",
    "transformers>=4.46.1",
]
